+++ src/main/java/joshie/harvest/buildings/HFBuildings.java
@@ -156,9 +156,11 @@ public class HFBuildings {
     //Configuration
     public static boolean FULL_BUILDING_RENDER;
     public static boolean CHEAT_BUILDINGS;
+    public static boolean ENABLE_DEMOLITION;
 
     public static void configure() {
         FULL_BUILDING_RENDER = getBoolean("Use Full Building Render", true);
         CHEAT_BUILDINGS = getBoolean("Use Cheat Buildings", false);
+        ENABLE_DEMOLITION = getBoolean("Enable demolition", true);
     }
 }