+++ src/main/java/joshie/harvest/buildings/placeable/entities/PlaceableItemFrame.java
@@ -2,12 +2,15 @@ package joshie.harvest.buildings.placeable.entities;
 
 import com.google.gson.annotations.Expose;
 import joshie.harvest.buildings.LootHelper;
+import joshie.harvest.core.helpers.EntityHelper;
+import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityHanging;
 import net.minecraft.entity.item.EntityItemFrame;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.Rotation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
@@ -25,6 +28,14 @@ public class PlaceableItemFrame extends PlaceableHanging {
         this.chestType = chestType;
         this.stack = stack;
         this.rotation = rotation;
+    }
+
+    @Override
+    public void remove(World world, BlockPos pos, Rotation rotation, ConstructionStage stage, IBlockState replacement) {
+        if (canPlace(stage)) {
+            BlockPos transformed = getTransformedPosition(pos, rotation);
+            EntityHelper.getEntities(EntityItemFrame.class, world, transformed, 0.5D, 0.5D).stream().forEach(Entity::setDead);
+        }
     }
 
     @Override