+++ src/main/java/joshie/harvest/buildings/placeable/entities/PlaceablePainting.java
@@ -1,10 +1,13 @@
 package joshie.harvest.buildings.placeable.entities;
 
 import com.google.gson.annotations.Expose;
+import joshie.harvest.core.helpers.EntityHelper;
+import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityHanging;
 import net.minecraft.entity.item.EntityPainting;
 import net.minecraft.util.EnumFacing;
+import net.minecraft.util.Rotation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
@@ -16,6 +19,14 @@ public class PlaceablePainting extends PlaceableHanging {
     public PlaceablePainting(String name, EnumFacing facing, int x, int y, int z) {
         super(facing, x, y, z);
         this.painting = name;
+    }
+
+    @Override
+    public void remove(World world, BlockPos pos, Rotation rotation, ConstructionStage stage, IBlockState replacement) {
+        if (canPlace(stage)) {
+            BlockPos transformed = getTransformedPosition(pos, rotation);
+            EntityHelper.getEntities(EntityPainting.class, world, transformed, 0.5D, 0.5D).stream().forEach(Entity::setDead);
+        }
     }
 
     @Override
@@ -35,7 +46,6 @@ public class PlaceablePainting extends PlaceableHanging {
     @Override
     public PlaceablePainting getCopyFromEntity(Entity e, int x, int y, int z) {
         EntityPainting p = (EntityPainting) e;
-
         return new PlaceablePainting(p.art.title, p.facingDirection, x, y, z);
     }
 }